requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.6
nvidia-ml-py3==7.352.0
prometheus-client==0.19.0
websockets==12.0
asyncio-mqtt==0.13.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0
dash==2.16.1
redis==5.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-json-logger==2.0.7
structlog==23.2.0
httpx==0.25.2
PyYAML==6.0.1 